add goBackBtn component and fixed sonar issues
